# solutions/Main_Server/swagger/swaggerDoc.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "IUM-TWEB Project 2023-2024 Query Documentation",
    "description": "This is the API documentation for the TWEB project for the years 2023-2024. The project is a football site that retrieves data from two servers, one running Express and the other running SpringBoot. Users can make queries to either of these servers or to the server running on port 3000, which contains all the routes for calls to the two servers."
  },
  "paths": {
...
    "/allCompetitions": {
      "get": {
        "summary": "Get all competitions",
        "description": "Retrieve a list of all competitions with ID, name, and country.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "Server Express - Main"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of competitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "competitionId": {
                        "type": "string",
                        "description": "Competition ID"
                      },
                      "name": {
                        "type": "string",
                        "description": "Competition name"
                      },
                      "countryName": {
                        "type": "string",
                        "description": "Country name"
                      }
                    }
                  },
                  "example": [
                    {
                      "competitionId": "CIT",
                      "name": "italy-cup",
                      "countryName": "Italy"
                    },
                    {
                      "competitionId": "NLSC",
                      "name": "johan-cruijff-schaal",
                      "countryName": "Netherlands"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Competitions not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Competitions"
        ]
      }
    },
...
    "/competitions": {
      "get": {
        "summary": "Get competitions by type",
        "description": "Returns a list of competitions for a given type. This endpoint is available only on the specified server.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "Server Express - Main"
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The type of competition"
          }
        ],
        "responses": {
          "200": {
            "description": "List of competitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "competitionId": {
                        "type": "string",
                        "description": "Competition ID"
                      },
                      "name": {
                        "type": "string",
                        "description": "Name of the competition"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No competitions found for this type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "No competitions found for this type"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "Bad request"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
...
  }
}